+++ vite-project/src/components/Loader.jsx
@@ -20,10 +20,11 @@
 // }
 
 // Loader.jsx
+import { Typography } from "@mui/material";
 import Box from "@mui/material/Box";
 import CircularProgress from "@mui/material/CircularProgress";
 
-export default function Loader() {
+export default function Loader({ loadingMessage }) {
   return (
     <Box
       sx={{
@@ -33,6 +34,7 @@ export default function Loader() {
         right: 0,
         bottom: 0,
         display: "flex",
+        flexDirection: "column",
         alignItems: "center",
         justifyContent: "center",
         backgroundColor: "rgba(0, 0, 0, 0.5)",
@@ -40,6 +42,7 @@ export default function Loader() {
       }}
     >
       <CircularProgress color="inherit" />
+      <Typography margin={2}>{loadingMessage}</Typography>
     </Box>
   );
 }
